requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-js-asset==2.1.0
django-mptt==0.14.0
django-shortcuts==1.6
djangorestframework==3.14.0
djangoshortcuts==0.0.5
Pillow==10.0.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
urllib3==1.26.16
